Split overview classes (#84)
* Extract OverviewTree to its own file and add unit tests
* Extract OverviewTreePager to its own file and add unit tests
* Remove old overview widgets tests
